Replace static FileHandler with depdency injected design
When book manager is created, it takes an instnace of an IFileHandler. The concrete implementation of this is given in the FileHandler class, and is created and injected in the Application class.

This design allows for mocking using Moq in unit tests. The BookManagerTests class has an example of this workflow.